requirements.txt
openai >= 1.1.1
python-dotenv >= 1.0.0
fastapi >= 0.94.1
pydantic >= 1.10.7
uvicorn >= 0.21.1
